requirements.txt
# ============================================
# Solana Trading Bot - Requirements
# ============================================

# -----------------
# Solana SDK
# -----------------
solana>=0.36.0                   # Solana Python SDK (RPC client, async API, websockets)
solders>=0.21.0                  # Low-level Solana types (Keypair, Pubkey, VersionedTransaction, etc.)

# -----------------
# HTTP Requests & Async
# -----------------
httpx>=0.27.0                    # Modern async HTTP client (for Jupiter API, DexScreener, etc.)
aiohttp>=3.9.0                   # Async HTTP library (backup/compatibility)

# -----------------
# Telegram Bot
# -----------------
python-telegram-bot>=22.0        # Telegram Bot API wrapper (Application, CommandHandler, etc.)

# -----------------
# Configuration & Settings
# -----------------
pyyaml>=6.0.1                    # YAML parser for config files
python-dotenv>=1.0.0             # .env file loading
pydantic>=2.5.0                  # Data validation and settings management
pydantic-settings>=2.1.0         # Pydantic settings integration

# -----------------
# Async Utilities
# -----------------
asyncio-throttle>=1.0.2          # Rate limiting for async operations

# -----------------
# Encoding & Cryptography
# -----------------
base58>=2.1.1                    # Base58 encoding/decoding for Solana keys

# -----------------
# Logging
# -----------------
structlog>=24.1.0                # Structured logging framework

# -----------------
# Removed GUI (Telegram Only)
# -----------------
# GUI removed - using Telegram bot interface only

# -----------------
# Development & Testing
# -----------------
pytest>=8.0.0                    # Testing framework
pytest-asyncio>=0.23.0           # Async test support
